requirements.txt
Flask==3.0.0
cpf-generator==2.1.4
names==0.3.0
python-dotenv==1.0.1
starkbank==2.23.0